PresentationBuilder moved to Clippit.PowerPoint namespace
