+++ OpenXmlPowerTools.Tests/OpenXmlPowerTools.Tests.csproj
@@ -4,7 +4,6 @@
     <TargetFrameworks>netcoreapp3.0</TargetFrameworks>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <LangVersion>8.0</LangVersion>
-    <!-- <Nullable>enable</Nullable> -->
     <GenerateBindingRedirectsOutputType>true</GenerateBindingRedirectsOutputType>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <RootNamespace>Clippit.Tests</RootNamespace>
