+++ OpenXmlPowerTools.Tests/PowerPoint/PresentationBuilderSlidePublishingTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Clippit.PowerPoint;
 using Xunit;
 
 namespace Clippit.Tests.PowerPoint
